Refactors Pair into SortablePair
SortablePair limits first element of pair to be Integer and implements Comparable interface
